requirements.txt
Flask-RESTful>=0.3.7,<=0.3.9
pytest>=5.2.1,<=5.2.9
Flask-SQLAlchemy>=2.4.1,<2.5.0
psycopg2>=2.8.4,<2.9.0
Flask-Migrate>=2.5.2,<2.6.0
bcrypt>=3.1.7,<3.2.0
passlib>=1.7.1,<1.8.0
flask-marshmallow>=0.10.1,<0.11.0
marshmallow-sqlalchemy>=0.19.0,<0.20.0